refactor(runner): hoist AbortRunnerOnDrop + summarize_actions to runner.rs [dirge-5oxu / R2]
The abort-on-drop guard was copy-pasted verbatim in review.rs and
phased_orchestrator.rs; the tool-action dedup-fold was duplicated 3x across
the review/curator passes. Hoist both to agent::runner (next to AgentRunner)
so the cancel-safety contract + the summary format each live in one place:

- AbortRunnerOnDrop (pub(crate)) — shared by the 4 review.rs forked-runner
  drains + phased_orchestrator's collect_runner_text. dirge-ba0m rationale
  preserved as a note at the review.rs call sites.
- summarize_actions(&[String]) — replaces the 3x identical
  fold(Vec<&str>).join(' · ') with one tested helper.

Behavior-preserving; per-site logging (targets/messages) left untouched.
Net -55 LOC. F4 (the 10-site spawn-and-assign 'launch_run' helper) is split
into its own PR — it spans the core UI loop, not just the vix code.

+ unit test for summarize_actions. 2452 tests pass; clean under -D warnings.
